Fist shot at implementing Aurafaction
The first try at implementing Aurafaction. I've implemented it as
replacement effect which triggers when a card is moved from the stack
to the graveyard without a null source. Hopefully this is a robust
enough solution. If it isn't, a check can be added directly, to
mage.game.stack.Spell.resolve, just like Bestow has. Some code starting
down that path I've left in a comment at the bottom of
AurafactionAbility.java.